Adds save and import functionality for Target
Adds the ability to serialize and deserialize the `Target` struct to and from JSON files.

This allows for saving and loading target profiles, enabling users to persist their work and resume it later. Implements `Saveable` and `Importable` traits with corresponding methods for JSON serialization and deserialization.
Also introduces `serde` crate with `Serialize` and `Deserialize` derives to Target struct.
Updates `Expandable` trait name to be more descriptive.